Staging

+++ apps/jobs/business_views.py
@@ -1,32 +1,37 @@
-from typing import Optional, Literal, List
+from datetime import datetime, time
+from typing import Literal, Optional, List
 from uuid import UUID
+
+import pytz
 from config.permissions import IsBusinessUser
 from django.db import transaction
 from django.db.models import Q
 from django.utils import timezone
 from helpers.utils import convert_base64_to_image_file
+from monkeypatches.response import Response
 from ninja import Router, PatchDict
 from ninja.errors import HttpError
-from ninja.responses import Response
-from ninja_extra.schemas import PaginatedResponseSchema
 from ninja_jwt.authentication import JWTAuth
 
-from accounts.models import Department, Role, SkillCategory
+from accounts.models import Department, Role, SkillCategory, Skill
+from accounts.schemas.talent import SkillSchema
+from chats.schemas import ResponseSchema
 from notification.notifications import send_talents_job_matching_notification
 from paginations import CustomPageNumberPaginationExtra
+from paginations import CustomPaginatedResponseSchema as PaginatedResponseSchema
 
-from monkeypatches.q_cluster import async_task
 from . import schemas as job_schemas
-from .enums import JobStatusType, PhaseType, QuestionTypeEnum
+from .enums import JobStatusType, PhaseType, QuestionTypeEnum, ActionType
 from .models import (
     EmploymentType, BusinessModel, JobLevel, JobPost, Job, RequiredAttribute, JobApplication, AvailableDay,
-    ScreeningQuestion, QuestionOption, Answer, Qualification
+    ScreeningQuestion, QuestionOption, Answer
 )
 from .schemas import (
-    EmploymentTypeSchema, CreateJobSchema, DepartmentSchema, RoleSchema, SkillCategorySchema, GenericNameAndUidSchema,
-    JobLevelSchema, TalentListJobPostSchema, JobDetailSchema, JobWorkflowViewPaginatedSchema
+    EmploymentTypeSchema, DepartmentSchema, RoleSchema, SkillCategorySchema, GenericNameAndUidSchema,
+    JobLevelSchema, BulkJobPostSchema, JobDetailSchema, JobWorkflowViewPaginatedSchema,
+    TalentListJobPostSchema
 )
-from .services import notify_business_on_matched_talents
+from .services import set_job_required_attributes, get_screening_questions_service
 
 router = Router(tags=["Business Jobs"])
 pagination_class = lambda page_size: CustomPageNumberPaginationExtra(page_size=page_size or 50)
@@ -64,20 +69,25 @@ def get_job_levels(request, search=""):
         queryset = queryset.filter(name__icontains=search)
     return queryset
 
-@router.get("qualifications", response=list[JobLevelSchema], tags=["Common"])
-def get_qualififcations(request, search=""):
-    queryset = Qualification.objects.all()
-    if search:
-        queryset = queryset.filter(name__icontains=search)
-    return queryset
-
 @router.get("skill-categories", response={200: list[SkillCategorySchema]}, tags=["Common"])
-def get_skills(request, search=""):
+def get_skills_categories(request, search="", category=""):
     queryset = SkillCategory.objects.all().prefetch_related("skill_set")
     if search:
         queryset = queryset.filter(Q(name__icontains=search)|
                                    Q(skill__name__icontains=search)).distinct("uid")
+    if category:
+        queryset = queryset.filter(name__iexact=category)
     return Response(data=[SkillCategorySchema.from_orm(q, context={"search": search}) for q in queryset])
+
+
+@router.get("skills", response={200: list[SkillSchema]}, tags=["Common"])
+def get_skills(request, search="", category=""):
+    queryset = Skill.objects.all()
+    if search:
+        queryset = queryset.filter(name__icontains=search)
+    if category:
+        queryset = queryset.filter(category__name__iexact=category)
+    return queryset.distinct("name").order_by("name")
 
 
 @router.get("business-models", response=list[GenericNameAndUidSchema], tags=["Common"])
@@ -88,19 +98,15 @@ def get_business_models(request, search=""):
     return queryset
 
 @router.patch("{job_uid}/required-attributes", response=job_schemas.RequiredAttributeSchema, auth=JWTAuth())
+@transaction.atomic
 def set_required_attributes(request, data:job_schemas.MutateRequiredAttributeSchema, job_uid:UUID):
     IsBusinessUser.check(request)
     business_user = request.user.businessuser
-    request_data = data.dict()
     job = Job.objects.filter(created_by__business=business_user.business, uid=job_uid).first()
     if not job:
         raise HttpError(404, "Job not found")
-    required_attributes, _ = RequiredAttribute.objects.get_or_create(job=job)
-    required_attributes.skills.set(request_data.pop("skills"))
-    required_attributes.business_models.set(request_data.pop("business_models"))
-    required_attributes.update(**request_data)
-    async_task(notify_business_on_matched_talents, job=job)
-    return required_attributes
+    request_data = data.dict()
+    return set_job_required_attributes(request_data, job)
 
 @router.get("{job_uid}/required-attributes", response=job_schemas.RequiredAttributeSchema, auth=JWTAuth())
 def get_required_attributes(request, job_uid:UUID):
@@ -121,10 +127,25 @@ def delete_job_post(request, job_post_uid:UUID):
     job_post =  JobPost.objects.filter(uid=job_post_uid, job__created_by__business=business_user.business).first()
     if not job_post:
         raise HttpError(404, "This job post does not exist")
+    if job_post.jobapplication_set.count() > 0:
+        raise HttpError(400, "Some job applications are tied to this job post")
     job_post.delete()
     return Response(status=204, data={"message": "Job post deleted"})
 
+@router.delete("job/{job_uid}", auth=JWTAuth())
+def delete_job(request, job_uid:UUID):
+    IsBusinessUser.check(request)
+    business_user = request.user.businessuser
+    job =  Job.objects.filter(uid=job_uid, created_by__business=business_user.business).first()
+    if not job:
+        raise HttpError(404, "This job does not exist")
+    if JobApplication.objects.filter(job_post__job=job).exists():
+        raise HttpError(400, "Some job applications are tied to this job")
+    job.delete()
+    return Response(status=204, data={"message": "Job deleted"})
+
 @router.patch("job-post/{job_post_uid}", response=job_schemas.JobPostDetailSchema, auth=JWTAuth())
+@transaction.atomic
 def update_job_post(request, job_post_uid, data: PatchDict[job_schemas.MutateJobPostSchema]):
     IsBusinessUser.check(request)
     business_user = request.user.businessuser
@@ -140,8 +161,22 @@ def update_job_post(request, job_post_uid, data: PatchDict[job_schemas.MutateJob
     job_post.update(**data)
     return job_post
 
+@router.patch("job-posts", response=ResponseSchema, auth=JWTAuth())
+@transaction.atomic
+def bulk_job_post_update(request, data: BulkJobPostSchema):
+    IsBusinessUser.check(request)
+    business_user = request.user.businessuser
+    job_posts = JobPost.objects.filter(job__created_by__business=business_user.business, uid__in=data.job_posts)
+    if data.action == ActionType.DELETE:
+        job_ids = (JobApplication.objects.filter(job_post__in=job_posts).
+                       only("job_post_id").values_list("job_post_id", flat=True))
+        job_posts.exclude(id__in=job_ids).delete()
+    else:
+        job_posts.update(status=data.action.value)
+    return Response({"message" : "actions have been applied successfully"}, status=200)
 
 @router.post("{job_uid}/job-post", response=job_schemas.JobPostDetailSchema, auth=JWTAuth())
+@transaction.atomic
 def add_job_post(request, job_uid:UUID, data: job_schemas.MutateJobPostSchema):
     IsBusinessUser.check(request)
     business_user = request.user.businessuser
@@ -153,7 +188,6 @@ def add_job_post(request, job_uid:UUID, data: job_schemas.MutateJobPostSchema):
         request_data["status"] = request_data["status"].value
         if request_data["status"] == JobStatusType.POSTED.value:
             request_data["posted_by"] = business_user
-            request_data["date_posted"] = timezone.now()
 
     job_post = JobPost(
         **request_data,
@@ -164,10 +198,13 @@ def add_job_post(request, job_uid:UUID, data: job_schemas.MutateJobPostSchema):
 
 
 @router.get("job-posts/{job_post_uid}", response=job_schemas.JobPostFullDetailSchema, auth=JWTAuth())
+@transaction.atomic
 def get_job_post_detail(request, job_post_uid):
     query = dict(uid=job_post_uid)
     if hasattr(request.user, "businessuser"):
         query["job__created_by__business"] = request.user.businessuser.business
+    if hasattr(request.user, "talent"):
+        request.context = dict(talent=request.user.talent)
     job_post = JobPost.objects.filter(**query).first()
     if not job_post:
         raise HttpError(404, "Job Post not found")
@@ -189,37 +226,43 @@ def get_talents_by_job_post(request, job_post_uid: UUID, search: str=None):
         )
     talents = job_post.get_talents()
     send_talents_job_matching_notification(talents.count(), job_post)
-    return talents.filter(query)
+    return talents.filter(query).order_by("-user__last_login")
 
 @router.post("", response=JobDetailSchema, auth=JWTAuth())
 @transaction.atomic
-def create_job(request, data:CreateJobSchema):
+def create_job(request, data:PatchDict[job_schemas.OptionalCreateJobSchema]):
     IsBusinessUser.check(request)
     business_user = request.user.businessuser
     business = business_user.business
-    request_data = data.dict()
-    request_data["created_by"] = business_user
-    availability = request_data.pop("availability")
-    screening_questions = request_data.pop("screening_questions")
-    job_posts = request_data.pop("job_posts")
-    additional_languages = request_data.pop("additional_languages", list())
-    skills = request_data.pop("skills", list())
-    business_models = request_data.pop("business_models", list())
+    data["created_by"] = business_user
+    availability = data.pop("availability", list())
+    screening_questions = data.pop("screening_questions", list())
+    required_attributes = data.pop("required_attributes", None)
+    job_posts = data.pop("job_posts", list())
+    additional_languages = data.pop("additional_languages", list())
+    skills = data.pop("skills", list())
+    business_models = data.pop("business_models", list())
 
-    if not request_data.get("hiring_company_name"):
-        request_data["hiring_company_name"] = business.name
-    if not request_data.get("hiring_company_description"):
-        request_data["hiring_company_description"] = business.description
-    request_data["work_structure"] = request_data["work_structure"].value
-    request_data["lunch_break"] = request_data["lunch_break"].value
-    request_data["technological_requirement"] = request_data["technological_requirement"].value
+    if not data.get("hiring_company_name"):
+        data["hiring_company_name"] = business.name
+    if not data.get("hiring_company_description"):
+        data["hiring_company_description"] = business.description
+    if "work_structure" in data:
+        data["work_structure"] = data["work_structure"].value
+    if "lunch_break" in data:
+        data["lunch_break"] = data["lunch_break"].value
+    if "technological_requirement" in data:
+        data["technological_requirement"] = data["technological_requirement"].value
 
 
-    job = Job.objects.create(**request_data)
+    job = Job.objects.create(**data)
     job.business_models.set(business_models)
     job.skills.set(skills)
     job.additional_languages.set(additional_languages)
     job.save()
+
+    if required_attributes:
+        set_job_required_attributes(required_attributes, job)
 
     for available_day in availability:
         available_day["day"] = available_day["day"].value
@@ -234,8 +277,9 @@ def create_job(request, data:CreateJobSchema):
             if job.availableday_set.filter(day=day).exists():
                 raise HttpError(400, f"{day} already exists")
             AvailableDay.objects.create(**available_day, job=job)
-    if request_data.get("logo"):
-        request_data["logo"] = convert_base64_to_image_file(request_data["logo"])
+    if data.get("logo"):
+        data["logo"] = convert_base64_to_image_file(data["logo"])
+
 
     for job_post in job_posts:
         job_post["status"] = job_post["status"].value if job_post.get("status") else JobStatusType.DRAFT.value
@@ -260,7 +304,7 @@ def update_job(request, data:PatchDict[job_schemas.UpdateJobSchema], job_uid:UUI
     additional_languages = data.pop("additional_languages", list())
     skills = data.pop("skills", list())
     business_models = data.pop("business_models", list())
-
+    required_attributes = data.pop("required_attributes", None)
     if not data.get("hiring_company_name"):
         data["hiring_company_name"] = business.name
     if not data.get("hiring_company_description"):
@@ -296,6 +340,8 @@ def update_job(request, data:PatchDict[job_schemas.UpdateJobSchema], job_uid:UUI
     if additional_languages:
         job.additional_languages.set(additional_languages)
     job.save()
+    if required_attributes:
+        set_job_required_attributes(required_attributes, job)
 
     return job
 
@@ -315,18 +361,18 @@ def job_list(request, page_size=50, page=1, search="", status:JobStatusType=None
 
     pagination = pagination_class(page_size).Input(page=page, page_size=page_size)
     return pagination_class(page_size).paginate_queryset(
-        queryset=queryset,
+        queryset=queryset.order_by("-created_at"),
         request=request,
         pagination=pagination,
         roles=queryset.count(),
         posts=business_user.business.job_posts().filter(job__in=queryset).count()
     )
 
-@router.get("{job_uid}", response=job_schemas.JobDetailSchema, auth=JWTAuth())
+@router.get("{job_uid}", response=job_schemas.FullJobDetailSchema, auth=JWTAuth())
 def job_detail(request, job_uid:UUID):
     IsBusinessUser.check(request)
     business_user = request.user.businessuser
-    job = Job.objects.filter(created_by__business=business_user.business, uid=job_uid).first()
+    job = Job.objects.prefetch_related("jobpost_set").filter(created_by__business=business_user.business, uid=job_uid).first()
     if not job:
         raise HttpError(404, "This job does not exist")
     return job
@@ -338,7 +384,8 @@ def view_applicants(request, job_post_uid:UUID, page_size=50, page=1, phase:Opti
 "match", "created_at", "stage", "phase", "experience"]]=None, asc:bool=True, search:str="", ):
     IsBusinessUser.check(request)
     business = request.user.businessuser.business
-    queryset = JobApplication.objects.filter(job_post__uid=job_post_uid, recruiter__business=business)
+    queryset = JobApplication.objects.filter(job_post__uid=job_post_uid, recruiter__business=business,
+                                             applicant__deleted_at__isnull=True)
     if search:
         queryset = queryset.filter(Q(
             Q(applicant__user__fullname__icontains=search)|
@@ -349,9 +396,9 @@ def view_applicants(request, job_post_uid:UUID, page_size=50, page=1, phase:Opti
         queryset = queryset.filter(stage__phase=phase.value)
     if new_application is not None:
         if new_application is True:
-            queryset = queryset.filter(stage__isnull=True)
+            queryset = queryset.filter(Q(stage__isnull=True)| Q(stage__phase=PhaseType.NEW.value))
         else:
-            queryset = queryset.filter(stage__isnull=False)
+            queryset = queryset.filter(Q(stage__isnull=False) & ~Q(stage__phase=PhaseType.NEW.value))
     if sort_by:
         sign = "-" if asc is False else ""
         if sort_by == "applicant":
@@ -374,6 +421,7 @@ def view_applicants(request, job_post_uid:UUID, page_size=50, page=1, phase:Opti
     )
 
 @router.patch("job-posts/applications/{application_uid}", response=job_schemas.JobApplicationListSchema, auth=JWTAuth())
+@transaction.atomic
 def update_application(request, application_uid:UUID, data:job_schemas.UpdateApplicationSchema):
     IsBusinessUser.check(request)
     application = JobApplication.objects.filter(uid=application_uid, recruiter__business=request.user.businessuser.business).first()
@@ -513,10 +561,17 @@ def delete_screening_question(request, question_uid:UUID):
 @router.get("{job_uid}/screening-questions", response=List[job_schemas.QuestionSchema], auth=JWTAuth(),
             tags=["Screening Test"])
 def get_screening_questions(request, job_uid: UUID):
-    screening_questions = ScreeningQuestion.objects.filter(job__uid=job_uid)
-    if hasattr(request.user, "businessuser"):
-        screening_questions = screening_questions.filter(job__created_by__business=request.user.businessuser.business)
-    return screening_questions
+    IsBusinessUser.check(request)
+    return get_screening_questions_service(request, job_uid)
+
+@router.get("{job_uid}/indeed/screener-questions",tags=["Screening Test"])
+def get_indeed_screening_questions(request, job_uid: UUID):
+    from services.job_posting.services.indeed import screening_question_to_indeed_screener_questions
+    job = Job.objects.filter(uid=job_uid).first()
+    if not job:
+        raise HttpError(404, "This job does not exist")
+    return screening_question_to_indeed_screener_questions(job)
+
 
 
 @router.get("applications/{application_uid}/screening-answers", response=List[job_schemas.ScreeningAnswerSchema], auth=JWTAuth(),
@@ -527,7 +582,5 @@ def get_screening_answers(request, application_uid:UUID):
     if not application:
         raise HttpError(404, "This application does not exist")
     return Answer.objects.filter(application=application)
-
-
